Create generate_souffled_playlist function in souffler entity
Move business logic out of souffle_playlist use case
Move existing souffler entity functions to souffler_util
Add Playlist namedtuple

As per PR #4
