+++ playlist_souffle/definitions/playlist.py
@@ -1,0 +1,5 @@
+"""Module for defining the Playlist namedtuple"""
+
+from collections import namedtuple
+
+Playlist = namedtuple('Playlist', 'owner name tracks description')
