+++ playlist_souffle/use_cases/souffle_playlist.py
@@ -1,12 +1,14 @@
 """Module for the souffle_playlist use case."""
 
 from datetime import datetime as dt
+from playlist_souffle.definitions.playlist import Playlist
 import playlist_souffle.entities.souffler as souffler_entity
 import playlist_souffle.entities.track as track_entity
 
 def souffle_playlist(spotify, playlist_uri, user_id, shuffle_by):
     """Souffle a playlist."""
 
+    playlist_name = spotify.fetch_playlist_name(playlist_uri)
     playlist_tracks = spotify.fetch_playlist_tracks(playlist_uri)
 
     collection_id_by_track = {track: track_entity.extract_track_collection_id(track, shuffle_by)
@@ -15,25 +17,21 @@ def souffle_playlist(spotify, playlist_uri, user_id, shuffle_by):
     collection_tracks_by_track = {track: spotify.fetch_collection_tracks(collection_id, shuffle_by)
                                   for track, collection_id in collection_id_by_track.items()}
 
-    souffled_playlist_tracks = souffler_entity.souffle_tracks(
-        playlist_tracks,
-        collection_tracks_by_track
+
+    playlist = Playlist(
+        owner=user_id,
+        name=playlist_name,
+        tracks=playlist_tracks,
+        description=''
     )
 
-    playlist_name = spotify.fetch_playlist_name(playlist_uri)
-
-    souffled_playlist_name = souffler_entity.generate_souffle_name(playlist_name)
-    souffled_playlist_description = souffler_entity.generate_souffle_description(
-        playlist_name,
+    souffled_playlist = souffler_entity.generate_souffled_playlist(
+        playlist,
         shuffle_by,
+        collection_tracks_by_track,
         dt.now()
     )
 
-    souffled_playlist_uri = spotify.create_playlist_with_tracks(
-        user_id,
-        souffled_playlist_name,
-        souffled_playlist_tracks,
-        description=souffled_playlist_description
-    )
+    souffled_playlist_uri = spotify.create_playlist(souffled_playlist)
 
     return souffled_playlist_uri
